Added ways of loading and displaying images.
The different methods are shown in comments. Either a Pixbuf is used, which represents a pixel buffer. Pixbuf can load images from the filesystem and scale them immediately or afterwards. The GtkImage widget is also capable of directly loading images, but doesn't support scaling.

The change to the compiler is recommended by GTK, for unknown reasons.