Fixed error where images would fail to load.
It was caused by the inotify IN_CREATE event, which is triggered as soon as a file is created. This caused the file to only be half written during file read, thus leading to an error. Now, the IN_CLOSE_WRITE event is used instead, which triggers only when a file is done writing.